Move VhostUserHandler to handler.rs
Move VhostUserHandler to handler.rs to ease maintenance.
